Create config and index.d to make index cleaner
